pyproject.toml
[tool.poetry]
name = "csd"
version = "0.1.0"
description = ""
authors = [ "Ben <benjamin@selas.studio>"]
readme = "README.md"
dependencies = {python = "^3.7", hivemind = "^1.1.0", grpcio="1.43.0", pydantic="1.9.2", numpy="^1.23.2", protobuf="^3.20.1", msgpack="^1.0.4", open-clip-torch = "^1.3.0", tsd = {git = "https://github.com/limiteinductive/tidy-stable-diffusion.git"}, tomli = "^2.0.1", ray = "^1.13.0"}


[build-system]
requires = [ "relaxed-poetry-core>=0.1.2"]
build-backend = "poetry.core.masonry.api"